moved fixture related classes to new package
