+++ livingdoc-engine/src/main/kotlin/org/livingdoc/engine/fixtures/FixtureFieldInjector.kt
@@ -1,4 +1,4 @@
-package org.livingdoc.engine.reflection
+package org.livingdoc.engine.fixtures
 
 import org.livingdoc.fixture.api.converter.TypeConverter
 import java.lang.reflect.Field
