+++ livingdoc-engine/src/main/kotlin/org/livingdoc/engine/fixtures/TypeConverterManager.kt
@@ -1,4 +1,4 @@
-package org.livingdoc.engine.reflection
+package org.livingdoc.engine.fixtures
 
 import org.livingdoc.fixture.api.converter.TypeConverter
 import org.slf4j.Logger
