+++ livingdoc-engine/src/main/kotlin/org/livingdoc/engine/fixtures/TypeConverters.kt
@@ -1,4 +1,4 @@
-package org.livingdoc.engine.reflection
+package org.livingdoc.engine.fixtures
 
 import org.livingdoc.fixture.api.converter.Converter
 import org.livingdoc.fixture.api.converter.TypeConverter
