+++ livingdoc-engine/src/test/kotlin/org/livingdoc/engine/fixtures/FixtureMethodInvokerTest.kt
@@ -1,4 +1,4 @@
-package org.livingdoc.engine.reflection
+package org.livingdoc.engine.fixtures
 
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Assertions.assertThrows
@@ -6,7 +6,7 @@ import org.junit.jupiter.api.Nested
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.ValueSource
-import org.livingdoc.engine.reflection.FixtureMethodInvoker.*
+import org.livingdoc.engine.fixtures.FixtureMethodInvoker.*
 import java.lang.reflect.Method
 import kotlin.reflect.KClass
